Use run number to generate unique file names in case of write
The write tests differ slightly from the setup logic used in the read
test case. In the latter, existing files are skipped as they can be
reused during the read tests.

This cannot be done for write performance tests, because this would
invalidate the measurements. Instead the tool panics as the file already
exists. This is caused by reusing the same thread id in multiple
successive runs, e.g. tests for thread counts 1 and 2 have the same
payload size and hostname and will therefore reuse the thread id of 1.

One possible solution is to remove the file directly after it has been
written. This causes consistency issues with some storages. Therefore
the run id is used to generate unique file names.