Fix Insecure option that broke the build
The option to trust all certificates wasn't fully implemented.
It broke the build.

As long as there is no cli param for this option we set it to "true" for
the S3 client.

The FS client doesn't need the Insecure option.
Region option is also not relevant for FS client.