Introduce sbmark.Report model
The Report model takes the aggregated results of a benchmark run.
It contains some metadata about the test run and the environment
as well as a list of Records, which itself contains the actual
throughput and latency percentiles of each test run.

The Report model can be used to render different output formats from.