Switched from csv upload to csv out file
The option -upload-csv was used to upload the final report to the test
bucket or folder. This is not very convenient because you would need to
download the .csv report after the test run.

The -upload-csv option is now simply called .csv.
It takes a simple name of a .csv file or a fully qualified, local path.

If set the benchmark results are saved in a local .csv file.