Add optional '-description' cli param
The description is part of the final report and currently only included
in the .json output. This is because I don't wanted to change the .csv
output.

The description should help to identify the test run and its purpose
later on. Additional informations about the client and server
environments are also part of the report (only .json) which helps to
understand and interpret the test result at a later point in time. Those
two are currently generated automatically and doesn't provide enough
information for a human reader.

The goal is to have the .json report a self included piece of
information.