Add latency stats for DNS, TCP, TLS and SRV processing
The TTLB (Time  to last byte) sums up the latencies of all necessary
steps of the protocol.

An S3 operation for example is a sequence of
1. DNS Lookup
2. TCP Handshake
3. TLS Handshake
4. Send HTTP Request
5. Server processing
6. Receive HTTP Response

This commit measures all the parts and adds latency avg, max, min and
percentiles to the .json report.

This way one can see where the time is lost when interpreting the total
TTLB latency.