Every sample uses an individual object
Formally the object names were individual per thread.
Each sample executes a single operation (like reading or writing an
object). So for example ...
- when using 4 samples and 1 thread the very same object would be
executed 4 times.
- when using 4 samples and 4 threads, each thread would execute
the operation on a differnt object.

This makes is harder to compare result for the different numbers of
threads because the system under test might behave differently.
For example because of caching effects when reexecuting the same
operation on the same object.

This commit makes sure that every sample is based on an individual
object so that runs with different number of threads are more
comparable.