Disable HTTP connection pooling to force TCP/TLS handshakes
We want to measure TCP and TLS handshakes in every sample, so that we
can calculate averages and percentiles.
At the same time we want to reuse the HTTP client throughout the
different go routines because else the server return HTTP 500 at a
certain number of HTTP connections.

The TCP/TLS caches can be disabled when HTTP connection pooling is
disabled in the transport of the HTTP client.