Add an option to transform old .json format to current format
The .json files that can be created with -json option contain a json
representation of the whole BenchmarkContext (incl. the BenchmarkReport
struct). The model is changing from version to version which makes it
harder to process it outside this application. Especially when
storage-benchmark .json files are stored in a central benchmark
database.

This change introduces downward compatibility of the BenchmarkContext
and BenchmarkReport models as well as an CLI option to transform old
.json files to the current version.