Add option to pretty print an existing json result to the console
When storage-benchmark is executing a pretty console output is printed
to the console. With the -json option a .json report is written to a
file. The .json reports are usually used to document and persist the
benchmark results in a formal file format. But the nice, humen readable
console output is usually not persisted, which is a pitty.

With this commit you now can regenerate the pretty console output
from an existing .json report for better human readability.

Usage:
storage-benchmark -print -json report.json